update: User update and delete

+++ controllers/UserController.js
@@ -51,8 +51,8 @@ class UserController {
    * @param {import('express').NextFunction} next
    */
   static async login(req, res, next) {
-    const { email, password } = req.body;
     try {
+      const { email, password } = req.body;
       const user = await User.findOne({ where: { email } });
       if (!user) throw { name: 'userNotFound' };
 
@@ -63,9 +63,8 @@ class UserController {
         id: user.id,
         email: user.email,
       });
-      res.status(200).json({ message: 'Success You are login', token });
+      res.status(200).json({ token });
     } catch (error) {
-      console.log(error);
       next(error);
     }
   }
@@ -76,17 +75,38 @@ class UserController {
    * @param {import('express').NextFunction} next
    */
   static async update(req, res, next) {
-    const { userId } = req.params;
-    const { email, full_name, username, profile_image_url, age, phone_number } =
-      req.body;
     try {
-      const user = await User.update(
-        { email, full_name, username, profile_image_url, age, phone_number },
-        { where: { id: userId }, returning: true }
-      );
-      res.status(200).json({ message: 'Success', data: user });
+      const { userId } = req.params;
+      // authorize
+      if (userId != req.user.id) throw { name: 'Forbidden' };
+
+      const requiredProps = [
+        'email',
+        'full_name',
+        'username',
+        'profile_image_url',
+        'age',
+        'phone_number',
+      ];
+      const body = {};
+      for (const prop of requiredProps) {
+        if (!req.body[prop]) throw { name: 'BadRequest' };
+        body[prop] = req.body[prop];
+      }
+
+      const userUpdated = await User.update(body, {
+        where: { id: userId },
+        returning: true,
+      });
+
+      const user = { ...userUpdated[1][0].dataValues };
+      delete user['id'];
+      delete user['password'];
+      delete user['createdAt'];
+      delete user['updatedAt'];
+
+      res.status(200).json({ user });
     } catch (error) {
-      console.log(error);
       next(error);
     }
   }
@@ -97,10 +117,15 @@ class UserController {
    * @param {import('express').NextFunction} next
    */
   static async delete(req, res, next) {
-    const { userId } = req.params;
     try {
-      const user = await User.destroy({ where: { id: userId } });
-      res.status(200).json({ message: 'Success Destroy' });
+      const { userId } = req.params;
+      // authorize
+      if (userId != req.user.id) throw { name: 'Forbidden' };
+
+      await User.destroy({ where: { id: userId } });
+      res
+        .status(200)
+        .json({ message: 'Your account has been successfully deleted' });
     } catch (error) {
       next(error);
     }
